divided logic game file and rspec into two different files per class. Added more negative test cases
